+++ robot_descriptions/_repositories.py
@@ -66,6 +66,11 @@ REPOSITORIES: Dict[str, Repository] = {
         commit="937f4186d6458bd682a7dae825fb6f4efe56ec69",
         cache_path="bhand_model",
     ),
+    "booster_gym": Repository(
+        url="https://github.com/BoosterRobotics/booster_gym.git",
+        commit="687a33d08b08875fe45dc8d91b54db83766df8b9",
+        cache_path="booster_gym",
+    ),
     "bullet3": Repository(
         url="https://github.com/bulletphysics/bullet3.git",
         commit="7dee3436e747958e7088dfdcea0e4ae031ce619e",
@@ -174,6 +179,14 @@ REPOSITORIES: Dict[str, Repository] = {
         url="https://github.com/mathieu-celerier/kinova_mj_description.git",
         commit="cee556b0a438e6904634a90826d4e8d2e005cd1f",
         cache_path="kinova_mj_description",
+    ),
+    "limxdynamics_robot-description": Repository(
+        url="https://github.com/limxdynamics/robot-description.git",
+        commit="a097533372a08298d45af391cbdfc2fd2dc3da6f",
+        # TODO: allow cache path with sub-directories; here we want the
+        # "robot_description" as last directory name, to match the package path
+        # in URDF descriptions so that Bullet loads them successfully
+        cache_path="robot_description",
     ),
     "mini_cheetah_urdf": Repository(
         url="https://github.com/Derek-TH-Wang/mini_cheetah_urdf.git",
@@ -312,7 +325,12 @@ REPOSITORIES: Dict[str, Repository] = {
     ),
     "Wiki-GRx-Models": Repository(
         url="https://github.com/FFTAI/Wiki-GRx-Models.git",
-        commit="f8e683f00d1d99deb882deb9dfce6030095b466a",
+        commit="351245ac8fa4bf6f4b0c41556e1e6976a438bcef",
         cache_path="Wiki-GRx-Models",
+    ),
+    "Wiki-GRx-Models-FourierN1": Repository(
+        url="https://github.com/FFTAI/Wiki-GRx-Models.git",
+        commit="f8e683f00d1d99deb882deb9dfce6030095b466a",  # FourierN1 branch
+        cache_path="Wiki-GRx-Models-FourierN1",
     ),
 }