+++ robot_descriptions/booster_t1_mj_description.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+#
+# SPDX-License-Identifier: Apache-2.0
+# Copyright 2025 Inria
 
-"""Fourier N1 description."""
+"""Booster T1 MJCF description."""
 
 from os import getenv as _getenv
 from os import path as _path
@@ -9,10 +12,10 @@ from os import path as _path
 from ._cache import clone_to_cache as _clone_to_cache
 
 REPOSITORY_PATH: str = _clone_to_cache(
-    "Wiki-GRx-Models-FourierN1",
+    "mujoco_menagerie",
     commit=_getenv("ROBOT_DESCRIPTION_COMMIT", None),
 )
 
-PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "N1")
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "booster_t1")
 
-URDF_PATH: str = _path.join(PACKAGE_PATH, "urdf", "N1_raw.urdf")
+MJCF_PATH: str = _path.join(PACKAGE_PATH, "t1.xml")