+++ CHANGELOG.md
@@ -8,8 +8,13 @@ All notable changes to this project will be documented in this file.
 
 - CLI: Add `pull` command to `robot_description` and cache it (thanks to @haixuantao)
 - Description: BamBot (URDF)
+- Description: Booster T1 (MJCF)
+- Description: Booster T1 (URDF)
+- Description: Fourier GR-1 (URDF)
 - Description: SO ARM 101 (MJCF) (thanks to @haixuantao)
 - Description: SO ARM 101 (URDF) (thanks to @haixuantao)
+- Description: WL P311D (URDF)
+- Description: WL P311E (URDF)
 
 ### Changed